requirements.txt
anthropic>=0.39.0
openai>=1.0.0
supabase>=2.0.0
numpy>=1.24.0
python-docx>=1.0.0
python-dotenv>=1.0.0
fastapi>=0.100.0
uvicorn>=0.23.0
pydantic>=2.0.0
python-multipart>=0.0.6
pdf2docx>=0.5.8
docx2pdf>=0.1.8